+++ app/models/User.scala
@@ -1,56 +1,13 @@
 package models
 
 import play.api.libs.json.{Format, Json}
+import models.{ReadBook, Collection}
 import reactivemongo.bson.{BSONArray, BSONDocument, BSONDocumentReader, BSONDocumentWriter}
-
-case class ReadBook(
-                     _id: String,
-                     title: String,
-                     author: String,
-                     genre: String,
-                     pages: Int,
-                     image: String,
-                     rating: Int,
-                     dateCompleted: String
-                   )
-
-object ReadBook {
-  implicit val fmt: Format[ReadBook] = Json.format[ReadBook]
-
-  implicit object ReadBookBSONReader extends BSONDocumentReader[ReadBook] {
-    def read(doc: BSONDocument): ReadBook = {
-      ReadBook(
-        doc.getAs[String]("_id").get,
-        doc.getAs[String]("title").get,
-        doc.getAs[String]("author").get,
-        doc.getAs[String]("genre").get,
-        doc.getAs[Int]("pages").get,
-        doc.getAs[String]("image").get,
-        doc.getAs[Int]("rating").get,
-        doc.getAs[String]("dateCompleted").get
-      )
-    }
-  }
-
-  implicit object ReadBookBSONWriter extends BSONDocumentWriter[ReadBook] {
-    def write(book: ReadBook): BSONDocument = {
-      BSONDocument(
-        "_id" -> book._id,
-        "title" -> book.title,
-        "author" -> book.author,
-        "genre" -> book.genre,
-        "pages" -> book.pages,
-        "image" -> book.image,
-        "rating" -> book.rating,
-        "dateCompleted" -> book.dateCompleted
-      )
-    }
-  }
-}
 
 case class User(
                  _id: String,
-                 readingHistory: List[ReadBook] = List()
+                 readingHistory: List[ReadBook] = List(),
+                 collections: List[Collection] = List(Collection("Future Reads"))
                )
 object User {
   implicit val fmt: Format[User] = Json.format[User]
@@ -61,6 +18,9 @@ object User {
         doc.getAs[String]("_id").get,
         doc.getAs[BSONArray]("readingHistory").get.toMap.toList.map(book => {
           ReadBook.ReadBookBSONReader.read(book._2.asInstanceOf[BSONDocument])
+        }),
+        doc.getAs[BSONArray]("collections").get.toMap.toList.map(collection => {
+          Collection.CollectionBSONReader.read(collection._2.asInstanceOf[BSONDocument])
         })
       )
     }
@@ -72,6 +32,9 @@ object User {
         "_id" -> user._id,
         "readingHistory" -> BSONArray(user.readingHistory.map {
           book => ReadBook.ReadBookBSONWriter.write(book)
+        }),
+        "collections" -> BSONArray(user.collections.map {
+          collection => Collection.CollectionBSONWriter.write(collection)
         })
       )
     }
